requirements.txt
Django>=5.1.5
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
PyMuPDF>=1.23.0
openai>=1.0.0
PyPDF2>=3.0.0
requests>=2.31.0
python-decouple>=3.8

